requirements.txt
openai>=1.0.0
pymongo>=4.0.0
requests>=2.28.0
python-dotenv>=1.0.0
tabulate>=0.8.9   